update and consolidate principal constants
